+++ metricflow/query/group_by_item/candidate_push_down/push_down_visitor.py
@@ -31,7 +31,7 @@ from metricflow.query.group_by_item.resolution_dag.resolution_nodes.query_resolu
     QueryGroupByItemResolutionNode,
 )
 from metricflow.query.group_by_item.resolution_path import MetricFlowQueryResolutionPath
-from metricflow.query.issues.group_by_item_resolver.no_common import NoCommonItemsInParents
+from metricflow.query.issues.group_by_item_resolver.no_common_items import NoCommonItemsInParents
 from metricflow.query.issues.group_by_item_resolver.no_matching_items_for_measure import NoMatchingItemsForMeasure
 from metricflow.query.issues.group_by_item_resolver.no_matching_items_for_no_metrics_query import (
     NoMatchingItemsForNoMetricsQuery,
